+++ simulant/utils/simplex.cpp
@@ -65,15 +65,15 @@ Simplex::Simplex(int seed):
 
 }
 
-float Simplex::noise(float x, float y) {
+float Simplex::noise(double x, double y) {
     return noise(x, y, 0, 0);
 }
 
-float Simplex::noise(float x, float y, float z) {
+float Simplex::noise(double x, double y, double z) {
     return noise(x, y, z, 0);
 }
 
-float Simplex::noise(float x, float y, float z, float w) {
+float Simplex::noise(double x, double y, double z, double w) {
     if(!initialized_) {
         init();
     }
